requirements.txt
google-cloud-bigquery
pandas
fsspec
gcsfs
pyarrow